server/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["wynardtage0 <fishface6666@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
alembic = "^1.13.1"
fastapi = "^0.111.0"
PyJWT = "^2.8.0"
python-dotenv = "^1.0.1"
passlib = "^1.7.4"
asyncpg = "^0.29.0"
uvicorn = "^0.29.0"
cryptography = "^42.0.7"
pydantic-async-validation = "^0.1.4"
pytz = "^2024.1"
sqlalchemy = "^2.0.30"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
